refactor create reservation to modal
