+++ src/components/CreateStudio.jsx
@@ -1,8 +1,6 @@
-const CreateStudio = () => {
-  return (
-    <div className={'modal-overlay show-modal modal-overlay'}>
-      CreateStudio Forrm
-    </div>
-  );
-};
+const CreateStudio = () => (
+  <div className="modal-overlay show-modal modal-overlay">
+    CreateStudio Forrm
+  </div>
+);
 export default CreateStudio;
